# modelindexer: Run active in dedicated goroutine