--- a/changelogs/head.asciidoc
+++ b/changelogs/head.asciidoc
@@ -22,3 +22,4 @@ Set error.id for OpenTelemetry exception span events {pull}9372[9372]
 ==== Added
 - Metrics data streams now use synthetic source {pull}9215[9215]
 - Service metrics now support global labels {pull}9248[9248]
+- Improve Elasticsearch output performance, particularly when compression is enabled (default) {pull}9318[9318]