--- a/internal/beater/beater.go
+++ b/internal/beater/beater.go
@@ -886,6 +886,8 @@ func (s *serverRunner) newFinalBatchProcessor(
 		stats := indexer.Stats()
 		v.OnKey("available")
 		v.OnInt(stats.AvailableBulkRequests)
+		v.OnKey("active")
+		v.OnInt(stats.ActiveBulkRequests)
 		v.OnKey("completed")
 		v.OnInt(stats.BulkRequests)
 	})