--- a/internal/beater/server_test.go
+++ b/internal/beater/server_test.go
@@ -975,8 +975,11 @@ func TestServerElasticsearchOutput(t *testing.T) {
 	assert.Equal(t, map[string]interface{}{
 		"elasticsearch": map[string]interface{}{
 			"bulk_requests": map[string]interface{}{
-				"available": int64(9),
-				"completed": int64(0),
+				"active":     int64(1),
+				"available":  int64(9),
+				"completed":  int64(0),
+				"upscales":   int64(0),
+				"downscales": int64(0),
 			},
 		},
 	}, snapshot)