--- a/internal/model/modelindexer/indexer_test.go
+++ b/internal/model/modelindexer/indexer_test.go
@@ -572,6 +572,7 @@ func TestModelIndexerCloseBusyIndexer(t *testing.T) {
 			Namespace: "testing",
 		}}}
 		err = indexer.ProcessBatch(context.Background(), &batch)
+		assert.NoError(t, err)
 	}
 
 	assert.NoError(t, indexer.Close(context.Background()))