@marclop 2022-10-10
I've included two commits of the implementation, the first one uses a local `sync.Mutex` in the active goroutine, not dissimilar to how the model indexer used locking previously (451e7f2).

```console
$ benchstat old.txt single.txt
name                               old time/op    new time/op    delta
ModelIndexer/NoCompression-8         1.11µs ± 1%    1.19µs ±15%   +7.61%  (p=0.008 n=5+5)
ModelIndexer/BestSpeed-8             2.03µs ± 3%    1.89µs ± 2%   -6.91%  (p=0.008 n=5+5)
ModelIndexer/DefaultCompression-8    4.10µs ± 2%    2.34µs ± 1%  -42.98%  (p=0.008 n=5+5)
ModelIndexer/BestCompression-8       6.95µs ± 1%    4.20µs ± 1%  -39.58%  (p=0.008 n=5+5)

name                               old alloc/op   new alloc/op   delta
ModelIndexer/NoCompression-8         2.07kB ± 0%    2.14kB ± 0%   +3.30%  (p=0.016 n=4+5)
ModelIndexer/BestSpeed-8             2.57kB ± 0%    2.56kB ± 0%   -0.35%  (p=0.008 n=5+5)
ModelIndexer/DefaultCompression-8    2.55kB ± 0%    2.52kB ± 0%   -0.83%  (p=0.008 n=5+5)
ModelIndexer/BestCompression-8       2.58kB ± 0%    2.56kB ± 0%   -0.62%  (p=0.008 n=5+5)

name                               old allocs/op  new allocs/op  delta
ModelIndexer/NoCompression-8           20.0 ± 0%      20.0 ± 0%     ~     (all equal)
ModelIndexer/BestSpeed-8               24.0 ± 0%      24.0 ± 0%     ~     (all equal)
ModelIndexer/DefaultCompression-8      24.0 ± 0%      24.0 ± 0%     ~     (all equal)
ModelIndexer/BestCompression-8         24.0 ± 0%      24.0 ± 0%     ~     (all equal)
```

The second one implements the same "algorithm" but without using channels instead of a mutex. It seems to have comparable performance: (ef72d77):

```console
$ benchstat old.txt single-lockless.txt
name                               old time/op    new time/op    delta
ModelIndexer/NoCompression-8         1.11µs ± 1%    1.15µs ± 4%   +4.41%  (p=0.008 n=5+5)
ModelIndexer/BestSpeed-8             2.03µs ± 3%    1.88µs ± 1%   -7.14%  (p=0.008 n=5+5)
ModelIndexer/DefaultCompression-8    4.10µs ± 2%    2.36µs ± 2%  -42.32%  (p=0.008 n=5+5)
ModelIndexer/BestCompression-8       6.95µs ± 1%    4.21µs ± 1%  -39.48%  (p=0.008 n=5+5)

name                               old alloc/op   new alloc/op   delta
ModelIndexer/NoCompression-8         2.07kB ± 0%    2.14kB ± 0%   +3.44%  (p=0.016 n=4+5)
ModelIndexer/BestSpeed-8             2.57kB ± 0%    2.56kB ± 0%   -0.31%  (p=0.008 n=5+5)
ModelIndexer/DefaultCompression-8    2.55kB ± 0%    2.52kB ± 0%   -0.87%  (p=0.008 n=5+5)
ModelIndexer/BestCompression-8       2.58kB ± 0%    2.57kB ± 0%   -0.52%  (p=0.008 n=5+5)

name                               old allocs/op  new allocs/op  delta
ModelIndexer/NoCompression-8           20.0 ± 0%      20.0 ± 0%     ~     (all equal)
ModelIndexer/BestSpeed-8               24.0 ± 0%      24.0 ± 0%     ~     (all equal)
ModelIndexer/DefaultCompression-8      24.0 ± 0%      24.0 ± 0%     ~     (all equal)
ModelIndexer/BestCompression-8         24.0 ± 0%      24.0 ± 0%     ~     (all equal)
```

```console
$ benchstat single.txt single-lockless.txt
name                               old time/op    new time/op    delta
ModelIndexer/NoCompression-8         1.19µs ±15%    1.15µs ± 4%   ~     (p=0.841 n=5+5)
ModelIndexer/BestSpeed-8             1.89µs ± 2%    1.88µs ± 1%   ~     (p=0.889 n=5+5)
ModelIndexer/DefaultCompression-8    2.34µs ± 1%    2.36µs ± 2%   ~     (p=0.548 n=5+5)
ModelIndexer/BestCompression-8       4.20µs ± 1%    4.21µs ± 1%   ~     (p=0.690 n=5+5)

name                               old alloc/op   new alloc/op   delta
ModelIndexer/NoCompression-8         2.14kB ± 0%    2.14kB ± 0%   ~     (p=0.579 n=5+5)
ModelIndexer/BestSpeed-8             2.56kB ± 0%    2.56kB ± 0%   ~     (p=0.667 n=5+5)
ModelIndexer/DefaultCompression-8    2.52kB ± 0%    2.52kB ± 0%   ~     (p=0.373 n=5+5)
ModelIndexer/BestCompression-8       2.56kB ± 0%    2.57kB ± 0%   ~     (p=0.905 n=5+5)

name                               old allocs/op  new allocs/op  delta
ModelIndexer/NoCompression-8           20.0 ± 0%      20.0 ± 0%   ~     (all equal)
ModelIndexer/BestSpeed-8               24.0 ± 0%      24.0 ± 0%   ~     (all equal)
ModelIndexer/DefaultCompression-8      24.0 ± 0%      24.0 ± 0%   ~     (all equal)
ModelIndexer/BestCompression-8         24.0 ± 0%      24.0 ± 0%   ~     (all equal)
```