@marclop 2022-10-10
I think the channel lockless approach has a race condition which causes the active indexer to never be flushed. I haven't managed to track down where/how it happens, so it may be best to leave that approach as a future optimization and update this PR to use the local mutex, which seems to be equal performance wise.