@marclop 2022-10-11
@axw benchmarks look good however I do notice an increase in overall RSS which could be problematic (and is) for 1GB APM Servers, and cause them to run out of memory.

Bear in mind that I have done this testing with a slightly smaller ES deployment (6 nodes instead of 12, totaling `116g`, rather than `232g`). Even then, we've got a ~20% performance increase in all sizes except `1gb` and  2GB` which is impressive.

## 2GB

```console
$ name          old time/op                  new time/op                  delta
AgentAll-128                   2.18s ± 0%                   2.13s ± 0%    -2.20%  (p=0.100 n=3+3)

name          old events/sec               new events/sec               delta
AgentAll-128                   8.28k ± 0%                   8.46k ± 0%    +2.21%  (p=0.100 n=3+3)

name          old max_goroutines           new max_goroutines           delta
AgentAll-128                     146 ± 8%                     198 ±21%   +35.70%  (p=0.100 n=3+3)

name          old max_heap_alloc           new max_heap_alloc           delta
AgentAll-128                    896M ± 1%                   1059M ± 6%   +18.24%  (p=0.100 n=3+3)

name          old max_heap_objects         new max_heap_objects         delta
AgentAll-128                   8.24M ± 1%                   9.49M ± 4%   +15.17%  (p=0.100 n=3+3)

name          old max_rss                  new max_rss                  delta
AgentAll-128                    965M ± 1%                   1167M ± 5%   +20.98%  (p=0.100 n=3+3)

name          old alloc/op                 new alloc/op                 delta
AgentAll-128                   558MB ± 0%                   560MB ± 0%    +0.37%  (p=0.100 n=3+3)

name          old allocs/op                new allocs/op                delta
AgentAll-128                   7.93M ± 0%                   7.94M ± 0%    +0.10%  (p=0.100 n=3+3)
```

## 4GB

```console
$ name          old time/op                  new time/op                  delta
AgentAll-256                   1.34s ± 1%                   1.12s ± 0%   -16.57%  (p=0.100 n=3+3)

name          old events/sec               new events/sec               delta
AgentAll-256                   13.4k ± 1%                   16.1k ± 0%   +19.87%  (p=0.100 n=3+3)

name          old max_goroutines           new max_goroutines           delta
AgentAll-256                     176 ± 8%                     250 ± 4%   +41.97%  (p=0.100 n=3+3)

name          old max_heap_alloc           new max_heap_alloc           delta
AgentAll-256                    914M ± 2%                   1096M ± 0%   +19.95%  (p=0.100 n=3+3)

name          old max_heap_objects         new max_heap_objects         delta
AgentAll-256                   8.14M ± 1%                   9.81M ± 1%   +20.46%  (p=0.100 n=3+3)

name          old max_rss                  new max_rss                  delta
AgentAll-256                    987M ± 1%                   1195M ± 1%   +21.10%  (p=0.100 n=3+3)

name          old mean_available_indexers  new mean_available_indexers  delta
AgentAll-256                    5.78 ± 1%                   19.82 ± 0%  +242.65%  (p=0.100 n=3+3)

name          old alloc/op                 new alloc/op                 delta
AgentAll-256                   566MB ± 0%                   560MB ± 0%    -1.06%  (p=0.100 n=3+3)

name          old allocs/op                new allocs/op                delta
AgentAll-256                   8.02M ± 0%                   7.94M ± 0%    -0.91%  (p=0.100 n=3+3)
```

## 8GB

```console
$ name          old time/op                  new time/op                  delta
AgentAll-512                   673ms ± 0%                   566ms ± 8%   -15.94%  (p=0.100 n=3+3)

name          old events/sec               new events/sec               delta
AgentAll-512                   26.8k ± 0%                   32.0k ± 8%   +19.26%  (p=0.100 n=3+3)

name          old max_goroutines           new max_goroutines           delta
AgentAll-512                     284 ± 1%                     434 ± 6%   +52.93%  (p=0.100 n=3+3)

name          old max_heap_alloc           new max_heap_alloc           delta
AgentAll-512                    936M ± 1%                   1190M ± 2%   +27.12%  (p=0.100 n=3+3)

name          old max_heap_objects         new max_heap_objects         delta
AgentAll-512                   7.86M ± 2%                  10.50M ± 2%   +33.57%  (p=0.100 n=3+3)

name          old max_rss                  new max_rss                  delta
AgentAll-512                   1.04G ± 2%                   1.31G ± 2%   +25.60%  (p=0.100 n=3+3)

name          old mean_available_indexers  new mean_available_indexers  delta
AgentAll-512                    2.40 ± 1%                    9.59 ±14%  +299.17%  (p=0.100 n=3+3)

name          old alloc/op                 new alloc/op                 delta
AgentAll-512                   562MB ± 0%                   574MB ± 0%    +2.25%  (p=0.100 n=3+3)

name          old allocs/op                new allocs/op                delta
AgentAll-512                   7.98M ± 0%                   8.06M ± 0%    +0.94%  (p=0.100 n=3+3)
```

In terms of CPU usage, the old indexer only used about 50% CPU on an 8GB APM Server, while the new one goes as high as 67%.

I think we may need to revisit the semaphore size, particularly for smaller sizes and perhaps size it according to the amount of memory available to APM Server.